rename select to bottom sheet field
