--- src/components/form/Select/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Select'
