+++ src/components/index.ts
@@ -10,7 +10,7 @@ export { default as DateInputField } from './form/DateInputField'
 export { default as DropdownField } from './form/DropdownField'
 export { default as InputField } from './form/InputField'
 export { default as Radio } from './form/Radio'
-export { default as Select } from './form/Select'
+export { default as BottomSheetField } from './form/BottomSheetField'
 export { default as Switch } from './form/Switch'
 export { default as TextArea } from './form/TextArea'
 export { default as TimeInputField } from './form/TimeInputField'
